Home, MyAccount & Login
The Home.php is literally just filled with some basic information as you asked. The MyAccount.php does also as asked, displays items being sold by users and items sold, the items sold doesn't appear yet since there is nothing within the database and I didn't want to mess about with it too much but if it's populated with appropriate data it should work as expected. I've also added a very minor change to the login just to allow for session variables between MyAccount and Login.